Update to Kubebuilder v4-alpha layout, move controllers to internal/controllers, main.go to cmd/main.go, update Vector, operator-sdk, helm, controller-gen, cert-manager and golangci-lint
